+++ pipeline/scripts/app-data-v2/lib/odpt/build-translations.ts
@@ -8,8 +8,8 @@ import type {
   OdptStation,
   OdptStationOrder,
   OdptStationTimetable,
-} from '../../../../types/odpt-train';
-import type { Provider } from '../../../../types/resource-common';
+} from '../../../../src/types/odpt-train';
+import type { Provider } from '../../../../src/types/resource-common';
 import { extractStationShortId } from './build-stops';
 import { getHeadsignFromDestination } from './build-timetable';